initial attempt to add welcome guide via plugin
