+++ packages/edit-post/README.md
@@ -502,6 +502,10 @@ _Type_
 
 -   `Object`
 
+### WelcomeGuide
+
+Undocumented declaration.
+
 <!-- END TOKEN(Autogenerated API docs) -->
 
 ## Contributing to this package
