+++ packages/edit-post/src/plugins/welcome-guide/image.js
