+++ packages/edit-post/src/plugins/welcome-guide/style.scss
