+++ packages/edit-post/src/plugins/welcome-guide/template.js
