+++ packages/edit-post/src/style.scss
@@ -23,7 +23,7 @@
 @import "./components/sidebar/template-summary/style.scss";
 @import "./components/text-editor/style.scss";
 @import "./components/visual-editor/style.scss";
-@import "./components/welcome-guide/style.scss";
+@import "./plugins/welcome-guide/style.scss";
 
 /**
  * Animations
